Parameter

Parameters

PaddyFieldAlgorithm.Parameter — Type.

This represents a parameter which can be optimized by the Paddy Field Algorithm

source

PaddyFieldAlgorithm.norm — Function.

norm(v, value)

Normalizes a Real value with the rules provided by Parameter v

source

PaddyFieldAlgorithm.unnorm — Function.

unnorm(v, value)

Inverses the normalization performed on a Real value provided by Parameter v

source

Base.Math.clamp — Function.

clamp(v, value)

Ensures that a Real value is an allowed value for Parameter v

source

PaddyFieldAlgorithm.init — Function.

init(v)

Initializes a random value for Parameter v

source

PaddyFieldAlgorithm.ParameterValue — Type.

A ParameterValue is a combination of a value(v) and a gaussian.

source

PaddyFieldAlgorithm.DefaultParameter — Type.

A Parameter that is not normalized, but is limited by bounds (limit)

source

PaddyFieldAlgorithm.RangeNormalizedParameter — Type.

A Parameter that is normalized to be in a linear range.

source

PaddyFieldAlgorithm.ZNormalizedParameter — Type.

A Parameter that is normalized to be normally distributed with a mean and standard deviation

source

PaddyFieldAlgorithm.seed — Function.

seed(param, value; scale, scaled)

Mutates a ParameterValue value using the rules defined by Parameter param

source