Regression

Public Interface

PaddyFieldAlgorithm.LinearSolution — Type.

LinearSolution contains two variables, a slope(m) and an intercept(b).

It is a regressive Solution which fits the following function: $y(x) = m*x + b$

source

PaddyFieldAlgorithm.PolynomialSolution — Type.

PolynomialSolution contains a variable number of coefficients.

It is a regressive Solution which fits the following function: $y(x) = c0 * x^0 + c1 * x^1 + c2 * x^2 + c3 * x^3 + ... + cn * x^n$

source

PaddyFieldAlgorithm.TrigonometricSolution — Type.

TrigonometricSolution contains a variable number of coefficients.

It is a regressive Solution which fits the following function: $y(x) = c0 * x^0 + c1 * x^1 + c2 * x^2 + c3 * x^3 + ... + cn * x^n$

source