Regression

Regression Solution

PaddyFieldAlgorithm.RegressionSolution — Type.

A RegressionSolution attempts to fit a function to a set of points where the fitness is defined by the Mean Squared Error between the predicted y and true y

$fitness(x,y) = -\sum_{i}^{length(x)}(f(x_i) - y_i)^2$

Where f is the Solution, x is a one dimensional input space, and y is the output space. Note the negative sign in the equation above which is necessary as the goal of the Paddy Field Algorithm is to maximize fitness.

source

PaddyFieldAlgorithm.LinearSolution — Type.

LinearSolution contains two variables, a slope(m) and an intercept(b).

It is a regressive Solution which fits the following function: $y(x) = m*x + b$

The distance between a LinearSolution and a second LinearSolution is $d(m_1,m_2,b_1,b_2) = (m_1 - m_2)^2 + (b_1 - b_2)^2$

source

PaddyFieldAlgorithm.PolynomialSolution — Type.

PolynomialSolution is a regressive Solution which fits the following function: $y(x) = \sum_{i=0}^{Degree}c_i x^i$

Degree is defined by the user to be a positive integer and the distance between a PolynomialSolution and a second PolynomialSolution is $d(c^1,c^2) = \sum_{i=0}^{Degree}(c_i^1 - c_i^2)^2$

source

PaddyFieldAlgorithm.TrigonometricSolution — Type.

TrigonometricSolution is a regressive Solution which fits the following function: $y(x) = c_0 + \sum_{j=1}^{Degree/2}(c_{2j} cos(2jπx) + c_{2j+1} sin(2jπx))$

Degree is defined by the user to be a positive integer and the distance between a TrigonometricSolution and a second TrigonometricSolution is $d(c^1,c^2) = (c_0^1 - c_0^2)^2 + \sum_{i=1}^{Degree}(c_i^1 - c_i^2)^2$

source